+++ src/components/accordion/index.ts
@@ -1,0 +1,5 @@
+export { AccordionGroup } from './AccordionGroup';
+export type { AccordionGroupProps } from './AccordionGroup/AccordionGroup.types';
+
+export { AccordionItem } from './AccordionItem';
+export type { AccordionItemProps, AccordionItemStyle } from './AccordionItem/AccordionItem.types';